worker: move host mocks
